:closed_lock_with_key: Make requests to API manager endpoint rather than directly to search API
