+++ README.md
@@ -51,16 +51,15 @@ environment.
 | Variable                           | Description                                                                                                                                                | Default                   | Required   |
 | :--------------------------------- | :--------------------------------------------------------------------------------------------------------------------------------------------------------- | :------------------------ | :--------- |
 | `ADOBE_TRACKING_URL`               | [Adobe Analytics](https://www.adobe.com/analytics/adobe-analytics.html) Dynamic Tag Management URL                                                         |                           | No         |
-| `AS_INDEX`                         | Name of the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis) index used by the application                | organisationlookup-dev    | Yes        |
-| `AS_API_KEY`                       | `api-key` for the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis)                                        |                           | Yes        |
-| `AS_SERVICE_NAME`                  | Name of the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis)                                              | nhsuksearchintne-dev      | Yes        |
-| `AS_VERSION`                       | Version of the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis)                                           | 2017-11-11                | Yes        |
 | `BASIC_AUTH`                       | An MD5 encrypted [htpasswd](https://httpd.apache.org/docs/2.4/misc/password_encryptions.html)                                                              | test:test                 | No         |
 | `HOTJAR_ANALYTICS_TRACKING_ID`     | [Hotjar](https://www.hotjar.com/) tracking id                                                                                                              |                           | No         |
 | `GOOGLE_ANALYTICS_TRACKING_ID`     | [Google Analytics](https://www.google.co.uk/analytics) property id                                                                                         |                           | No         |
 | `LOG_LEVEL`                        | Numeric [log level](https://github.com/trentm/node-bunyan#levels)                                                                                          | Depends on `NODE_ENV`     | No         |
 | `NODE_ENV`                         | Node environment                                                                                                                                           | development               | Yes        |
 | `PORT`                             | Server port                                                                                                                                                | 3000                      | Yes        |
+| `SEARCH_API_KEY`                   | `api-key` for the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis)                                        |                           | Yes        |
+| `SEARCH_API_HOST`                      | Name of the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis)                                              | nhsuksearchintne-dev      | Yes        |
+| `SEARCH_API_VERSION`                   | Version of the [Azure Search Service](https://docs.microsoft.com/en-us/rest/api/searchservice/#calling-the-apis)                                           | 2017-11-11                | Yes        |
 | `WEBTRENDS_ANALYTICS_TRACKING_ID`  | [Webtrends](https://www.webtrends.com/) tracking id                                                                                                        |                           | No         |
 
 ## FAQ
