+++ app.js
@@ -4,7 +4,7 @@ const log = require('./app/lib/logger');
 const app = require('./server');
 const applicationStarts = require('./app/lib/prometheus/counters').applicationStarts;
 
-requireEnv(['AS_API_KEY']);
+requireEnv(['SEARCH_API_KEY']);
 
 app.listen(app.port, () => {
   applicationStarts.inc(1);
