--- app/lib/azuresearch/headers.js
@@ -1,6 +1,0 @@
-const asAPIKey = require('../../../config/config').azureSearch.key;
-
-module.exports = {
-  'Content-Type': 'application/json',
-  'api-key': asAPIKey,
-};
