+++ app/lib/search/headers.js
@@ -1,0 +1,6 @@
+const key = require('../../../config/config').search.apiKey;
+
+module.exports = {
+  'Content-Type': 'application/json',
+  'subscription-key': key,
+};
