+++ app/lib/search/queryBuilder.js
