+++ app/lib/search/request.js
@@ -1,0 +1,16 @@
+const rp = require('request-promise-native');
+
+const search = require('../../../config/config').search;
+const headers = require('./headers');
+
+async function request(query) {
+  return rp({
+    body: JSON.stringify(query),
+    headers,
+    method: 'POST',
+    // url: `https://${asConfig.serviceName}.search.windows.net/indexes/${asConfig.index}/docs/search?api-version=${asConfig.version}`,
+    url: `https://${search.host}/service-search/search?api-version=${search.version}`,
+  });
+}
+
+module.exports = request;
