+++ docker-compose-tests.yml
@@ -38,7 +38,7 @@ services:
     volumes:
       - sh-test-ci-node_modules:/code/node_modules
     environment:
-      AS_API_KEY: ${AS_API_KEY}
+      SEARCH_API_KEY: ${SEARCH_API_KEY}
       COVERALLS_REPO_TOKEN: ${COVERALLS_REPO_TOKEN}
       TRAVIS: ${TRAVIS}
       TRAVIS_BRANCH: ${TRAVIS_BRANCH}
