+++ docker-compose.yml
@@ -9,11 +9,10 @@ services:
         NODE_ENV: ${NODE_ENV-development}
     command: yarn run start-watch
     environment:
-      AS_API_KEY: ${AS_API_KEY}
-      AS_INDEX: ${AS_INDEX}
-      AS_SERVICE_NAME: ${AS_SERVICE_NAME}
-      AS_VERSION: ${AS_VERSION}
       NODE_ENV: ${NODE_ENV-development}
+      SEARCH_API_HOST: ${SEARCH_API_HOST}
+      SEARCH_API_KEY: ${SEARCH_API_KEY}
+      SEARCH_API_VERSION: ${SEARCH_API_VERSION}
 
   frontend:
     extends:
