+++ rancher-config/docker-compose.yml
@@ -6,13 +6,12 @@ services:
     image: "nhsuk/sexual-health-service-finder:${DOCKER_IMAGE_TAG}"
     environment:
       ADOBE_TRACKING_URL: ${ADOBE_TRACKING_URL}
-      AS_API_KEY: ${AS_API_KEY}
-      AS_INDEX: ${AS_INDEX}
-      AS_SERVICE_NAME: ${AS_SERVICE_NAME}
-      AS_VERSION: ${AS_VERSION}
       GOOGLE_ANALYTICS_TRACKING_ID: ${GOOGLE_ANALYTICS_TRACKING_ID}
       HOTJAR_ANALYTICS_TRACKING_ID: ${HOTJAR_ANALYTICS_TRACKING_ID}
       NODE_ENV: production
+      SEARCH_API_HOST: ${SEARCH_API_HOST}
+      SEARCH_API_KEY: ${SEARCH_API_KEY}
+      SEARCH_API_VERSION: ${SEARCH_API_VERSION}
       WEBTRENDS_ANALYTICS_TRACKING_ID: ${WEBTRENDS_ANALYTICS_TRACKING_ID}
     labels:
       io.rancher.container.pull_image: always
