+++ test/integration/resultsPage.js
@@ -3,13 +3,12 @@ const chaiHttp = require('chai-http');
 const cheerio = require('cheerio');
 
 const app = require('../../server');
-const asConfig = require('../../config/config').azureSearch;
 const constants = require('../../app/lib/constants');
 const getQueryType = require('../../app/lib/utils/queryMapper').getQueryType;
 const getTextOnlyFromElement = require('../lib/utils').getTextOnlyFromElement;
 const iExpect = require('../lib/expectations');
 const nockRequests = require('../lib/nockRequests');
-const queryBuilder = require('../../app/lib/azuresearch/queryBuilder');
+const queryBuilder = require('../../app/lib/search/queryBuilder');
 
 const expect = chai.expect;
 
@@ -48,7 +47,7 @@ describe('Results page results', () => {
       const type = queryParams.type;
 
       before(`make request for type:${type} and origin:${origin}`, async () => {
-        const path = `/indexes/${asConfig.index}/docs/search`;
+        const path = '/service-search/search';
         const searchOrigin = { location: { lat: 53.7974737203539, lon: -1.55262247079646 } };
         const queryType = getQueryType(type, origin);
         const query = queryBuilder(searchOrigin, queryType, 30);
@@ -114,8 +113,8 @@ describe('Results page results', () => {
     it('should display results when a location crosses 2 countries, one of which is England', async () => {
       const type = constants.serviceTypes.kit;
       const origin = constants.serviceChoices.over25;
-      const path = `/indexes/${asConfig.index}/docs/search`;
       const latLon = { location: { lat: 55.3977217554393, lon: -2.77657929395506 } };
+      const path = '/service-search/search';
       const queryType = getQueryType(type, origin);
       const query = queryBuilder(latLon, queryType, 30);
       const requestBody = JSON.stringify(query);
